代码调整

+++ azz-system-parent/azz-order-service/src/main/java/com/azz/selection/mapper/ClientSelectionRecordMapper.java
@@ -10,13 +10,13 @@ import com.azz.order.selection.vo.SelectionRecord;
 
 @Mapper
 public interface ClientSelectionRecordMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(ClientSelectionRecord record);
 
     int insertSelective(ClientSelectionRecord record);
 
-    ClientSelectionRecord selectByPrimaryKey(Integer id);
+    ClientSelectionRecord selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(ClientSelectionRecord record);
